pyproject.toml
[tool.poetry]
name = "emotion-mlops"
version = "0.1.0"
description = "MLOps пайплайн для классификации эмоций"
authors = ["albert <shoma.albert081205@gmail.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = ">=3.10,<3.15"
fastapi = "^0.121.0"
uvicorn = "^0.38.0"
pandas = "^2.3.3"
scikit-learn = "^1.7.2"
mlflow = "2.14.3"
pyyaml = "^6.0.3"
nltk = "^3.9.2"
joblib = "^1.5.2"
graphene = "2.1.9"
graphql-core = "2.3.2"
graphql-relay = "2.0.1"
setuptools = "^80.9.0"
torch = "^2.9.1"
transformers = "^4.57.3"
accelerate = ">=0.26.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
